workshop/frugal_alignment/requirements.txt
# Python 3.10+ required for unsloth
torch>=2.0.0
transformers>=4.40.0
datasets
peft
trl
bitsandbytes
accelerate
wandb
pyyaml

# Unsloth - requires Python 3.10+
# Install with: pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# unsloth @ git+https://github.com/unslothai/unsloth.git
